alt_point: only changing the alt_point values instead of original

+++ srcs/rotate_point.c
@@ -1,21 +1,22 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        ::::::::            */
-/*   point_rotations.c                                  :+:    :+:            */
+/*   rotate_point.c                                     :+:    :+:            */
 /*                                                     +:+                    */
 /*   By: mminkjan <mminkjan@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/12/07 11:20:20 by mminkjan       #+#    #+#                */
-/*   Updated: 2019/12/07 11:27:37 by mminkjan      ########   odam.nl         */
+/*   Updated: 2019/12/09 17:44:34 by mminkjan      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/fdf.h"
 
-void	iso_projection(t_points *alt_point)
+void	iso_projection(t_points *alt_point)	
 {
-	double		prev_x;
-	double		prev_y;
+	int		prev_x;
+	int		prev_y;
+
 
 	prev_x = alt_point->x;
 	prev_y = alt_point->y;
